@@ -11,8 +11,8 @@
 		}
 	],
 	"require": {
-		"php": ">=7.2.0",
-		"able/statics": "1.0.*",
+		"php": ">=7.4.0",
+		"able/statics": "dev-master",
 		"able/helpers": "dev-master",
 		"able/reglib": "dev-master",
 		"able/struct": "dev-master"
